Progress commit for saving state of potato game. See detail below.
As it turns out, the state of the potato game contains timers. Timers are not
serializable in the Gob format we use for state files, because they contain
channels. I forgot about that when I started working on this -- and I got pretty
far on it. This commit has the machinery for setting and updating state on
startup, as well as resetting timers. However, I really didn't want to go the
extra distance of separating out timers (really, we should just have one) from
the game state. So even though this is nearly complete, I'm giving up on it for
now -- I just don't think the benefit of persisted games outweighs the
complexity. But, I'm committing this so that I can come back to it later if I
change my mind.